Remove MIPS architecture
The MIPS architecture port has been added 5+ years ago in order to
support a Chrome OS project that ended up going nowhere. No other board
has used it since and nobody is still willing or has the expertise and
hardware to maintain it. We have decided that it has become too much of
a mainenance burden and the chance of anyone ever reviving it seems too
slim at this point. This patch eliminates all MIPS code and
MIPS-specific hacks.

Reviewed-on: https://review.coreboot.org/c/coreboot/+/34919
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/drivers/spi/cbfs_spi.c
@@ -53,7 +53,7 @@ static ssize_t spi_readat(const struct region_device *rd, void *b,
 		u64 speed;	/* KiB/s */
 		int bps;	/* Bits per second */
 
-		speed = size * 1000 / usecs;
+		speed = size * (u64)1000 / usecs;
 		bps = speed * 8;
 
 		printk(BIOS_DEBUG, "read SPI %#zx %#zx: %ld us, %lld KB/s, %d.%03d Mbps\n",
